code/requirements.txt
pytelegraf
ping3
requests
